Rename AdvancedAdbClient to AdbClient (AdvancedAdbClient is too long to remember)
