+++ AdvancedSharpAdbClient/Extensions/Factories.cs
@@ -24,10 +24,10 @@ namespace AdvancedSharpAdbClient
         public static Func<EndPoint, IAdbSocket> AdbSocketFactory { get; set; }
 
         /// <summary>
-        /// Gets or sets a delegate that creates a new instance of the <see cref="AdvancedAdbClient"/> class.
+        /// Gets or sets a delegate that creates a new instance of the <see cref="AdbClient"/> class.
         /// </summary>
-        /// <returns>A new instance of the <see cref="AdvancedAdbClient"/> class.</returns>
-        public static Func<EndPoint, IAdvancedAdbClient> AdbClientFactory { get; set; }
+        /// <returns>A new instance of the <see cref="AdbClient"/> class.</returns>
+        public static Func<EndPoint, IAdbClient> AdbClientFactory { get; set; }
 
         /// <summary>
         /// Gets or sets a function that returns a new instance of a class that implements the
@@ -41,7 +41,7 @@ namespace AdvancedSharpAdbClient
         /// <see cref="ISyncService"/> interface, that can be used to transfer files to and from
         /// a given device.
         /// </summary>
-        public static Func<IAdvancedAdbClient, DeviceData, ISyncService> SyncServiceFactory { get; set; }
+        public static Func<IAdbClient, DeviceData, ISyncService> SyncServiceFactory { get; set; }
 
         /// <summary>
         /// Resets all factories to their default values.
@@ -49,7 +49,7 @@ namespace AdvancedSharpAdbClient
         public static void Reset()
         {
             AdbSocketFactory = (endPoint) => new AdbSocket(endPoint);
-            AdbClientFactory = (endPoint) => new AdvancedAdbClient(endPoint, Factories.AdbSocketFactory);
+            AdbClientFactory = (endPoint) => new AdbClient(endPoint, Factories.AdbSocketFactory);
             AdbCommandLineClientFactory = (path) => new AdbCommandLineClient(path);
             SyncServiceFactory = (client, device) => new SyncService(client, device);
         }
