+++ AdvancedSharpAdbClient/Receivers/NamespaceDoc.cs
@@ -10,7 +10,7 @@ namespace AdvancedSharpAdbClient.Receivers
     /// <summary>
     /// <para>
     ///     This namespace contains various receivers that are able to parse Android console output. You can use
-    ///     the receivers in combination with the <see cref="AdvancedAdbClient.ExecuteRemoteCommandAsync(string, DeviceData, IShellOutputReceiver, CancellationToken)"/>
+    ///     the receivers in combination with the <see cref="AdbClient.ExecuteRemoteCommandAsync(string, DeviceData, IShellOutputReceiver, CancellationToken)"/>
     ///     method to capture the output of any Android command.
     /// </para>
     /// </summary>
